Router routes property needs to be in the prototype
At least for Backbone 1.1.2, Backbone.Router 's routes property needs to be in the prototype for it to work. So it must be a method or the property needs to be defined with a getter for it to work. I think this is the same as for View 's events property which is defined as method here.